Upload version 0.3.0
Add bower's installation;
Add possibility to set class of all multi-bar's blocks and not only
background color;
Add the field `shadow` in the options to set the shadow of your
multi-bar;
Add the field `title` in the options of the legend to set a title to
your multi-bar's legend;
Add style for legend. Now you can set the classes of the title and the
items of your multi-bar's legend.